+++ librina/src/librina-ipc-manager.cc
@@ -797,10 +797,10 @@ IPCProcess * IPCProcessFactory::create(
 		const ApplicationProcessNamingInformation& ipcProcessName,
 		const std::string& difType) throw (CreateIPCProcessException) {
 	lock();
-	int ipcProcessId = 1;
+	unsigned short ipcProcessId = 1;
 	unsigned int portId = 0;
 	pid_t pid=0;
-	for (int i = 1; i < 1000; i++) {
+	for (unsigned short i = 1; i < 1000; i++) {
 		if (ipcProcesses.find(i) == ipcProcesses.end()) {
 			ipcProcessId = i;
 			break;
@@ -880,13 +880,13 @@ IPCProcess * IPCProcessFactory::create(
 	return ipcProcess;
 }
 
-void IPCProcessFactory::destroy(unsigned int ipcProcessId)
+void IPCProcessFactory::destroy(unsigned short ipcProcessId)
 throw (DestroyIPCProcessException) {
 	lock();
 
 	int resultUserSpace = 0;
 	int resultKernel = 0;
-	std::map<int, IPCProcess*>::iterator iterator;
+	std::map<unsigned short, IPCProcess*>::iterator iterator;
 	iterator = ipcProcesses.find(ipcProcessId);
 	if (iterator == ipcProcesses.end())
 	{
@@ -927,7 +927,7 @@ std::vector<IPCProcess *> IPCProcessFactory::listIPCProcesses() {
 	std::vector<IPCProcess *> response;
 
 	lock();
-	for (std::map<int, IPCProcess*>::iterator it = ipcProcesses.begin();
+	for (std::map<unsigned short, IPCProcess*>::iterator it = ipcProcesses.begin();
 			it != ipcProcesses.end(); ++it) {
 		response.push_back(it->second);
 	}
@@ -936,10 +936,10 @@ std::vector<IPCProcess *> IPCProcessFactory::listIPCProcesses() {
 	return response;
 }
 
-IPCProcess * IPCProcessFactory::getIPCProcess(unsigned int ipcProcessId)
+IPCProcess * IPCProcessFactory::getIPCProcess(unsigned short ipcProcessId)
         throw (GetIPCProcessException)
 {
-        std::map<int, IPCProcess*>::iterator iterator;
+        std::map<unsigned short, IPCProcess*>::iterator iterator;
 
         lock();
         iterator = ipcProcesses.find(ipcProcessId);
